requirements_py36.txt
_tflow_select==2.3.0                       
absl-py==0.15.0             
aiohttp==3.7.4.post0      
astor==0.8.1            
async-timeout==3.0.1            
attrs==21.4.0             
blas==1.0                         
blinker==1.4              
brotlipy==0.7.0           
ca-certificates==2024.3.11            
cachetools==4.2.2              
certifi==2021.5.30        
cffi==1.14.6           
chardet==4.0.0           
charset-normalizer==2.0.4              
click==8.0.3              
coverage==5.5              
cryptography==35.0.0           
cython==0.29.24          
freetype==2.12.1               
gast==0.2.2                    
giflib==5.2.1                
google-auth==2.6.0              
google-auth-oauthlib==0.4.1                      
google-pasta==0.2.0              
grpcio==1.36.1           
h5py==2.10.0           
hdf5==1.10.4               
icc_rt==2022.1.0             
idna==3.3                
idna_ssl==1.1.0            
imageio==2.9.0              
importlib-metadata==4.8.1            
intel-openmp==2019.4                      
jpeg==9e                   
keras-applications==1.0.8                      
keras-preprocessing==1.1.2              
lerc==3.0                  
libdeflate==1.17                 
libmklml==2019.0.5             
libpng==1.6.39               
libprotobuf==3.17.2               
libtiff==4.5.1                
libwebp==1.3.2                
libwebp-base==1.3.2                
lz4-c==1.9.4                
markdown==3.3.4            
mkl==2020.2                      
mkl-service==2.3.0            
mkl_fft==1.3.0            
mkl_random==1.1.1            
multidict==5.1.0            
numpy==1.19.2           
numpy-base==1.19.2           
oauthlib==3.1.0                      
olefile==0.46               
opencv-python==4.5.3.56 
openssl==1.1.1w               
opt_einsum==3.3.0              
pillow==8.3.1            
pip==21.3.1                   
protobuf==3.17.2           
pyasn1==0.4.8              
pyasn1-modules==0.2.8                      
pycparser==2.21               
pyjwt==1.7.1                    
pyopenssl==22.0.0             
pyreadline==2.1                      
pysocks==1.7.1            
python==3.6.13               
requests==2.27.1             
requests-oauthlib==1.3.0                      
rsa==4.7.2              
scipy==1.5.2            
setuptools==59.6.0                   
six==1.16.0             
sqlite==3.41.2               
tensorboard==2.6.0                      
tensorboard-data-server==0.6.0            
tensorboard-plugin-wit==1.6.0                      
tensorflow==1.15.0          
tensorflow-base==1.15.0          
tensorflow-estimator==2.6.0              
termcolor==1.1.0            
tk==8.6.12               
typing-extensions==4.1.1                
typing_extensions==4.1.1              
urllib3==1.26.8             
vc==14.2                 
vs2015_runtime==14.27.29016          
werkzeug==0.16.1                     
wheel==0.37.1             
win_inet_pton==1.1.0            
wincertstore==0.2              
wrapt==1.12.1           
xz==5.4.6                
yarl==1.6.3            
zipp==3.6.0              
zlib==1.2.13               
zstd==1.5.5
Jinja2-3.0.3
MarkupSafe-2.0.1
Werkzeug-2.0.3
click-8.0.4
colorama-0.4.5
dataclasses-0.8
flask-2.0.3
importlib-metadata-4.8.3
itsdangerous-2.0.1
typing-extensions-4.1.1
zipp-3.6.0